Long to int fixed for id

+++ src/main/java/abl/frd/qremit/converter/nafex/repository/ApiT24ModelRepository.java
@@ -1,12 +1,10 @@
 package abl.frd.qremit.converter.nafex.repository;
-
-import abl.frd.qremit.converter.nafex.model.ApiBeftnModel;
 import abl.frd.qremit.converter.nafex.model.ApiT24Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ApiT24ModelRepository extends JpaRepository<ApiT24Model, Long> {
+public interface ApiT24ModelRepository extends JpaRepository<ApiT24Model, Integer> {
     ApiT24Model findByTransactionNo(String transactionNo);
     Optional<ApiT24Model> findByTransactionNoEqualsIgnoreCase(String transactionNo);
 }
